requirements.txt
Django>=4.2
djangorestframework>=3.14
drf-spectacular>=0.27
django-filter>=23.1
psycopg2-binary>=2.9
